first flush then change track
If it's done in opposite way, we sometimes loose caps event
in sinks.

If we have audio-tracks with different codecs, say
we have AAC and AC3. AAC is selected and running, we
switch to AC3, no caps event with AAC caps in sink ->
AC3 is written to decoder as AAC -> no sound.